backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pillow==9.5.0
tensorflow==2.16.2
numpy==1.26.4
opencv-python-headless==4.10.0.84

# Disc identification dependencies
transformers==4.37.2
torch==2.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# OCR dependencies
pytesseract==0.3.10
easyocr==1.7.0
paddlepaddle==2.6.2
paddleocr==2.8.1
